+++ src/engine/lighting/index.ts
@@ -1,7 +1,6 @@
 export * from './lighting-config';
 export type * from './flicker-options';
 export * from './darkness-component';
-export * from './ambient-light-component';
 export * from './point-light-component';
 export * from './cone-light-component';
 export * from './light-occluder-component';